Identify the default network interface of the node
For now, this is only implemented for Linux. It reads the file specified
by the new config option NetworkRouteFile, defaults to `/proc/net/route`
and it retrieves the name of the first interface it founds that has a
destination of '0'.

The value is cached using the integration caching mechanism and uses the
ttl defined with DiscoveryCacheTTL.