Node network metrics from default interface (#13)
* Identify the default network interface of the node

For now, this is only implemented for Linux. It reads the file specified
by the new config option NetworkRouteFile, defaults to `/proc/net/route`
and it retrieves the name of the first interface it founds that has a
destination of '0'.

The value is cached using the integration caching mechanism and uses the
ttl defined with DiscoveryCacheTTL.

* Network metrics fallback to default network interface

When querying the summary endpoint from Kubelet to get the Node or Pod
network metrics, if the default network interface is not eth0 then summary
endpoint for Kubelet doesn't return the metrics as we expect them. We rely on
them being a direct member of the "network" object. See rxBytes, txBytes and
rxErrors in the following example metrics:

```
"network": {
 "time": "2020-06-04T10:01:15Z",
 "name": "eth0",
 "rxBytes": 207909096,
 "rxErrors": 0,
 "txBytes": 8970981,
 "txErrors": 0,
 "interfaces": [
  {
   "name": "eth0",
   "rxBytes": 207909096,
   "rxErrors": 0,
   "txBytes": 8970981,
   "txErrors": 0
  },
  {
   "name": "ip6tnl0",
   "rxBytes": 0,
   "rxErrors": 0,
   "txBytes": 0,
   "txErrors": 0
  },
  {
   "name": "tunl0",
   "rxBytes": 0,
   "rxErrors": 0,
   "txBytes": 0,
   "txErrors": 0
  }
 ]
}
```

This scenario only happens when the default interface is eth0. Kubernetes
source code has it hardcoded that eth0 is the default. In the following
example you can see that we only have network metrics inside the interfaces
list, in this case there is no eth0 on the and the default interface is ens5:

```
"network": {
 "time": "2020-06-04T10:01:15Z",
 "name": "",
 "interfaces": [
  {
   "name": "ens5",
   "rxBytes": 207909096,
   "rxErrors": 42,
   "txBytes": 8970981,
   "txErrors": 24
  },
  {
   "name": "ip6tnl0",
   "rxBytes": 0,
   "rxErrors": 0,
   "txBytes": 0,
   "txErrors": 0
  },
  {
   "name": "tunl0",
   "rxBytes": 0,
   "rxErrors": 0,
   "txBytes": 0,
   "txErrors": 0
  }
 ]
 ```

In cases like this, the integration will look for the default interface
inside the interfaces list and use those values. The default interface name
is retrieved from the network route file (default /proc/net/route).

When running the unprivileged version of the integration we don't have access
to the route file, the integration won't be able to get the default interface
name and won't send network metrics for the unless there's a network
interface called eth0.

For Pods, this issue is mainly present when using hostNetwok since they
shared the same network interfaces with the Node.

* Add 1.18 cluster to e2e tests

For the E2E tests we use a jsonschema depending on the scenario, this are the exceptional cases:

- Kubernetes version 1.18 or newer with unprivileged, the node network
  metrics are optional. This is because only if the node has the eth0
  interface we will get the network information.

- K8bernetes version 1.18 or newer, the pod network metrics are
  optional. This is because pods using hostNetwork won't get the default
  interface metrics. We could improve this by finding out what pods use
  the hostNetwok in the test and load the schema dynamically for pods.

* Update changelog