discovery/ksm: improvements to discovery strategy (#104)
* discovery/ksm: fix usage of Trim for stripping suffix

* discovery/ksm: reorder label names being searched
In order to discover the pod where KSM is located, we make non-namespaced requests to the API server searching for different labels. The helm chart we're currently deploying as a dependency uses:
app.kubernetes.io/name=kube-state-metrics
This commit moves this label to be the first one being tried, which should reduce the number of requests to the API server to 1/3 when KSM is deployed with its helm chart (which is, to our understanding, the most common scenario).