NR-139168: Fix k8s.container.cpuCoresUtilization metric calculation
Background (as observed by folks on slack thread - https://newrelic.slack.com/archives/C043X7A8JRF/p1687808604387229): The 'cpuCoresUtilization' metric is a computed metric = 'cpuUsedCores' / 'cpuLimitCores'.

Further the 'cpuUsedCores' is another computed metric = 'usageNanoCores' / 10e9.

The 'usageNanoCores' is arbitrarily observed to hold a super high value driving up the 'cpuCoresUtilization' percentage value. Issue seems to be reproducible a few times per day on the cluster 'kb-aup-aws-us-east-1.pool.clarovideo.net'.

Analysis: The 'usageNanoCores' metric is scraped by the nri-agent's kubelet's pod both at the pod level and container level from the kubelet's `/metrics/summary` endpoint. I reviewed the `kubelet` open-source repo for more details and found this.

The kubelet currently gathers container metrics from either of the 2 sources - cAdvisor or Container-Runtime Interface (CRI). The CRI based metrics have to be enabled explicitly via the `PodAndContainerStatsFromCRI` kubernetes feature-gate. It is unlikely that the customer has enabled this.

The CPU UsageNanoCores metric is a computed metric and is cumulative. For cAdvisor please refer to the following code references to see how it is calculated.

https://github.com/kubernetes/kubernetes/blob/master/pkg/kubelet/stats/helper.go#L52

https://github.com/kubernetes/kubernetes/blob/master/vendor/github.com/google/cadvisor/info/v2/conversion.go#L276

https://github.com/kubernetes/kubernetes/blob/master/vendor/github.com/google/cadvisor/info/v2/conversion.go#L252

To summarize, the cAdvisor first checks if the currentValue is > lastValue and if it is true, it divides the difference with the time delta between current timestamp and last timestamp. Its hard to say precisely what could have caused this high value from the above calculation.

I tried searching for issues reported on the 'kubelet' repo and found this - https://github.com/kubernetes/kubernetes/issues/114057. Unfortunately the issue is closed as there were no owners assigned. I see a comment towards the end reporting that the issue was observed on v1.25 as well, so I don't suppose the cluster version upgrade will fix the above issue.

Suggested fix: I believe as a fix on nri-kubernetes integration, we may treat the scraped 'UsageNanoCores' metric value as invalid if the value is arbitrarily large. This fix would also serve as an additional check to ensure the value we ingest into NR is within the reasonable bounds.

I ran the below NRQL queries to observe the correlation between `cpuUsedCores (aka usageNanoCores)`, `cpuLimitCores` and `cpuCoresUtilization` where:

cpuCoresUtilization = (cpuUsedCores/cpuLimitCores) * 100

From the results, the observation was that the absurd cpuUsedCores numbers begin when :

cpuUsedCores >  cpuLimitCores * 10

So a reasonable assumption could be something like this:

cpuUsedCores >  cpuLimitCores * 100